refactor: split library models into separate files
