+++ library_system/library/models/author.py
@@ -1,0 +1,17 @@
+from django.db import models
+
+
+class Author(models.Model):
+    first_name = models.CharField(max_length=100, null=True, blank=True)
+    middle_name = models.CharField(max_length=100, null=True, blank=True)
+    last_name = models.CharField(max_length=100, null=True, blank=True)
+    date_of_birth = models.DateField(null=True, blank=True)
+    date_of_death = models.DateField(null=True, blank=True)
+
+    def __str__(self):
+        return self.full_name or f"{self.first_name} {self.last_name}"
+
+    @property
+    def full_name(self) -> str:
+        parts = [self.first_name, self.middle_name, self.last_name]
+        return " ".join(part for part in parts if part)
